requirements.txt
flask
pandas
psutil
PyYAML
pynvml